crewai_gcp/requirements.txt
crewai[tools]>=0.130.0,<1.0.0
google-cloud-aiplatform>=1.101.0
gradio>=5.35.0
fastmcp>=0.1.0
